moved logout component to auth
